pyproject.toml
[project]
name = "industrial-qa-backend"
version = "0.1.0"
description = "Industrial QA agent backend with FastAPI + RAG pipeline"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Industrial QA Team", email = "ops@example.com" }]
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "pydantic[email]>=2.8.0",
    "sqlalchemy>=2.0.30",
    "sqlalchemy[asyncio]>=2.0.30",
    "aiomysql>=0.2.0",
    "cryptography>=41.0.0",
    "alembic>=1.13.2",
    "httpx>=0.27.0",
    "qdrant-client>=1.9.0",
    "chromadb>=0.5.0",
    "rank-bm25>=0.2.2",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-openai>=0.1.0",
    "llama-index>=0.10.30",
    "openai>=1.35.0",
    "tiktoken>=0.7.0",
    "typer>=0.12.3",
    "python-dotenv>=1.0.1",
    "jinja2>=3.1.4",
    "bcrypt>=4.1.2",
    "pyjwt>=2.8.0",
    "redis>=5.0.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "unstructured>=0.11.0",
    "python-multipart>=0.0.6"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.7",
    "httpx>=0.27.0",
    "ruff>=0.4.8"
]

[tool.ruff]
# 行长度限制
line-length = 100

# 选择要启用的规则集
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (导入排序)
    "N",   # pep8-naming (命名规范)
    "UP",  # pyupgrade (Python版本升级)
    "B",   # flake8-bugbear (常见bug)
    "C4",  # flake8-comprehensions (推导式优化)
    "SIM", # flake8-simplify (代码简化)
]

# 忽略的规则
ignore = [
    "E501",  # 行长度（由formatter处理）
]

# 排除的目录
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.format]
# 代码格式化配置
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["Doc*", "docs*", "alembic*", "scripts*", "tests*"]

